Move createCSV firebase function into separate file
